+++ app/_app/immutable/entry/start.CBfJciUe.js
@@ -1,0 +1,1 @@
+import{a as r}from"../chunks/BQcA_xL3.js";import{x as t}from"../chunks/BUuoSTUH.js";export{t as load_css,r as start};